pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Анатолий Зайнутдинов <zainutdinov.aaa@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
